Clean up code and reorganize structure
