+++ core/models.py
@@ -1,13 +1,16 @@
 from enum import Enum
 
+
 class SendType(Enum):
     Feedback = 10
     Solution = 20
+
 
 class Complexity(Enum):
     Easy = 10
     Middle = 20
     Hard = 30
+
 
 complexityDict = {
     Complexity.Easy: "Легкий",
@@ -15,9 +18,10 @@ complexityDict = {
     Complexity.Hard: "Тяжелый",
 }
 
+
 class Task(object):
     def __init__(
-        self, number, name, announcement_link, solution_link=None, level=None, tags=None
+            self, number, name, announcement_link, solution_link=None, level=None, tags=None
     ):
         self.number = number
         self.name = name
@@ -29,9 +33,10 @@ class Task(object):
     def get_level(self):
         return complexityDict[self.level]
 
+
 class Puzzle(object):
     def __init__(
-        self, number, name, telegram_link, announcement_link, solution_link=None, tags=None
+            self, number, name, telegram_link, announcement_link, solution_link=None, tags=None
     ):
         self.number = number
         self.name = name
@@ -40,15 +45,16 @@ class Puzzle(object):
         self.solution_link = solution_link
         self.tags = tags
 
+
 class Event(object):
     def __init__(
-        self, number, name, description, date, link, type, location=None, time=None
+            self, number, name, description, date, link, _type, location=None, time=None
     ):
         self.number = number
         self.name = name
         self.description = description
         self.date = date
         self.link = link
-        self.type = type
+        self.type = _type
         self.location = location
         self.time = time
