+++ services/__init__.py
@@ -1,0 +1,4 @@
+from event import *
+from info import *
+from puzzle import *
+from task import *
