+++ services/task.py
@@ -1,10 +1,12 @@
-import requests
-from config import *
-from models import *
+from config import TASKS_URL, TASKS_URL_VERSION, DATA_ACCESS_KEY
+from core.models import Task
+
 from requests.structures import CaseInsensitiveDict
+import requests
 
 headers = CaseInsensitiveDict()
 headers["X-ACCESS-KEY"] = DATA_ACCESS_KEY
+
 
 class TaskService:
     def __init__(self, data=None, version=None):
